+++ __tests__/common.js
@@ -12,5 +12,5 @@ const transform = (code, opts) =>
 
 module.exports = {
   getCode: (source, opts) => transform(source, opts),
-  getBody: (source, opts) => transform(source, opts).split('\n', 2)[1],
+  getBody: (source, opts) => transform(source, opts).split('\n').slice(1).join(''),
 }