+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+# 1.1.0 - April 6, 2020
+
+- string template support
+- simplify args passed to helper
+
 # 1.0.0 - April 6, 2020
 
-* first version
-
+- first version